requirements.txt


# General pip dependencies
# aiohappyeyeballs==2.4.4
# aiohttp==3.11.10
# aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
blosc==1.11.2
git+https://github.com/openai/CLIP
diffusers==0.31.0
einops==0.8.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
h5py==3.12.1
huggingface-hub==0.26.5
importlib-metadata==8.5.0
lightning-utilities==0.11.9
multidict==6.1.0
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
propcache==0.2.1
pydantic==2.10.3
pydantic-core==2.27.1
python-dateutil==2.9.0.post0
pytorch-lightning
pytz==2024.2
regex==2024.11.6
safetensors==0.4.5
six==1.17.0
tokenizers==0.21.0
torchmetrics==1.6.0
transformers==4.47.0
typing-extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
yarl==1.18.3
zipp==3.21.0



### lotus
lxml==5.3.0
jupyterlab==4.2.5
notebook==6.1.5
absl_py==2.1.0
accelerate==0.30.1
addict==2.4.0
easydict==1.13
filelock==3.16.1
flash_attn==2.5.9.post1
Flask==3.0.3
groq==0.11.0
gym==0.26.2
hydra-core==1.3.2
jsonlines==4.0.0
lmdb==1.4.1
matplotlib==3.9.2
msgpack_numpy==0.4.8
msgpack_python==0.5.6
numpy==1.26.4
omegaconf==2.3.0
open3d==0.18.0
open_clip_torch==2.24.0
opencv_python==4.10.0.82
Pillow==10.4.0
pytest==8.3.3
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
scipy==1.14.1
setuptools==72.1.0

tensorboardX==2.6.2.2
termcolor==2.5.0
timm==1.0.9
tqdm==4.66.4

typed_argument_parser==1.10.0
yacs==0.1.8
spconv-cu118
typed-argument-parser
